update discourse group sync to work with latest API version

+++ SyncDiscourseGroups.py
@@ -28,12 +28,12 @@ class ChildScript(Script):
                               "Scholarship (12 months)"]
         self.group_name = 'MakeICT_Members'
         self.member_group_id = \
-            self.discourse_api.group(self.group_name)['group']['id']
+            self.discourse_api._get("/groups/{0}.json".format(self.group_name))['group']['id']
 
     def SyncUser(self, d_user, wa_contacts):
         # print(self.discourse_api.user_emails(d_user['user']['username']))
-        print(d_user['user']['username'], ':',
-              d_user['user']['emails'], ':',
+        print(d_user['username'], ':',
+              d_user['emails'], ':',
               [c['Email'] for c in wa_contacts])
         is_member = False
         is_active = False
@@ -58,35 +58,42 @@ class ChildScript(Script):
             print("User is an active member, adding user to group")
 
             try:
-                if d_user['user']['primary_group_name'] == self.group_name:
+                if d_user['primary_group_id'] == self.member_group_id:
                     print("User is already in group")
                     return
+                else:
+                    print("User is not in group")
             except KeyError:
                 print("User has no primary group")
 
             try:
                 self.discourse_api.add_group_member(
-                    self.member_group_id, d_user['user']['username'])
+                    self.member_group_id, d_user['username'])
             except DiscourseClientError:
                 print("Failed to add user to group")
 
         else:
-            print("user is not an active member, removing user from group")
-            self.discourse_api.delete_group_member(
-                self.member_group_id, d_user['user']['id'])
+            print("user is not an active member")
+            if(self.member_group_id == d_user['primary_group_id']):
+                print("removing user from group")
+                self.discourse_api.delete_group_member(
+                    self.member_group_id, d_user['id'])
+            else:
+                print("user is not in group")
 
     def SyncUsers(self, disc_users, wa_users):
         for u in disc_users:
             # if u['user']['username'] != '':
             #     continue
-            emails = self.discourse_api.user_emails(u['user']['username'])
+            u = self.discourse_api.user_all(u['user']['id'])
+            emails = self.discourse_api.user_emails(u['username'])
             email_list = [emails['email']]
             email_list += emails['secondary_emails']
-            u['user']['emails'] = email_list
+            u['emails'] = email_list
 
             found_contacts = []
             for contact in wa_users:
-                if contact['Email'] in u['user']['emails']:
+                if contact['Email'] in u['emails']:
                     print("Found WA user for %s" % contact['Email'])
                     found_contacts.append(contact)
 
@@ -95,7 +102,14 @@ class ChildScript(Script):
             time.sleep(2)
 
     def SyncAllUsers(self):
-        discourse_users = self.discourse_api.user_list()
+        response = self.discourse_api._get('/directory_items.json?period=all&order=post_count')
+        discourse_users = response['directory_items']
+        while response['meta']['total_rows_directory_items'] > len(discourse_users):
+            response = self.discourse_api._get(response['meta']['load_more_directory_items'])
+            discourse_users += response['directory_items']
+            print(response['meta']['total_rows_directory_items'])
+            print(len(discourse_users))
+
         wa_contacts = self.WA_API.GetAllContacts()
 
         self.SyncUsers(discourse_users, wa_contacts)
